+++ src/main/java/uk/q3c/krail/core/option/OptionPermission.java
@@ -16,6 +16,7 @@ package uk.q3c.krail.core.option;
 import org.apache.shiro.authz.permission.WildcardPermission;
 import uk.q3c.krail.core.shiro.SubjectIdentifier;
 import uk.q3c.krail.option.Option;
+import uk.q3c.krail.option.OptionEditAction;
 import uk.q3c.krail.option.OptionKey;
 import uk.q3c.krail.option.UserHierarchy;
 
@@ -44,9 +45,7 @@ import static com.google.common.base.Preconditions.*;
  */
 public class OptionPermission extends WildcardPermission {
 
-    public enum Action {
-        EDIT, VIEW
-    }
+
 
     private final String permissionString;
 
@@ -55,7 +54,7 @@ public class OptionPermission extends WildcardPermission {
      *
      * @param hierarchy
      */
-    public OptionPermission(Action action, UserHierarchy hierarchy, int index, OptionKey optionKey, String userId) {
+    public OptionPermission(OptionEditAction action, UserHierarchy hierarchy, int index, OptionKey optionKey, String userId) {
         checkNotNull(action);
         checkNotNull(hierarchy);
         checkNotNull(optionKey);