+++ src/main/java/uk/q3c/krail/option/OptionEditAction.java
@@ -1,0 +1,8 @@
+package uk.q3c.krail.option;
+
+/**
+ * Created by David Sowerby on 10 Aug 2017
+ */
+public enum OptionEditAction {
+    EDIT, VIEW
+}